return 401/403 & WWW-Authenticate by MagpieAdapter + provide Max-Age value in cookie
