+++ config/magpie.ini
@@ -25,18 +25,29 @@ pyramid.includes =
     ziggurat_foundations.ext.pyramid.get_user
 
 ## magpie
-# see 'Configuration' section of documentation for all options
+# see all options in 'https://pavics-magpie.readthedocs.io/en/latest/configuration.html'
 magpie.port = 2001
 magpie.url = http://localhost:2001
-magpie.max_restart = 5
+
+# magpie.config_path =
+
+# --- cookie definition --- (defaults below if omitted)
+# refer to 'https://pavics-magpie.readthedocs.io/en/latest/configuration.html#security-settings'
 # WARNING:
 #   Secret should be the same in Twitcher INI
 #   Otherwise, they cannot resolve same user references
 magpie.secret =
-magpie.push_phoenix = true
-# magpie.config_path =
+# magpie.cookie_name = auth_tkt
+# magpie.cookie_expire =    # value in seconds, never expire by default
 
-# caching settings refer to the Performance section in the documentation
+# --- temporary token definition --- (defaults below if omitted)
+magpie.token_expire = 86400
+
+# --- phoenix ---
+magpie.push_phoenix = true
+
+# --- caching settings ---
+# refer to 'https://pavics-magpie.readthedocs.io/en/latest/performance.html'
 # WARNING:
 #   Caching settings should also be defined in Twitcher INI as it is its application that will
 #   receive most requests and actually ask Magpie to resolve ACL/Services for it to allow/deny access
