+++ docs/authentication.rst
@@ -2,7 +2,7 @@
 
 .. _auth_methods:
 
-Authentication and Authorization Methods
+Authentication and Authorization
 ==========================================
 
 In order to perform :term:`Authentication` in `Magpie`, multiple :term:`Providers` and methods are supported.
@@ -180,26 +180,40 @@ Authentication Headers
 ---------------------------
 
 .. versionadded:: 3.8
-
     The ``WWW-Authentication`` and ``Location-When-Unauthenticated`` headers are returned whenever the
     HTTP ``Unauthorized [401]`` response is the result of a request. This is done in order to help requesting
     users or applications identify the endpoint where it can attempt :term:`Authentication` with credentials.
 
 
 After execution of an :term:`Authentication` request, a ``Set-Cookie`` header with `Magpie` user identification token
-named according to :ref:`config_security` should be set in the response. Web browsers and libraries for HTTP requests
-handling should automatically detect that header and register the ``Cookie`` for subsequent requests. Alternatively,
-the ``Cookie`` can be provided directly in the request using the following format.
+named according to :ref:`config_security` should be set in the response as follows.
 
 .. code-block::
 
-    {MAGPIE_COOKIE_NAME}=<auth-token>!userid_type:int; [Domain=<domain>; Path=<path>; HttpOnly; SameSite=Lax]
+    Set-Cookie: {MAGPIE_COOKIE_NAME}=<auth-token>!userid_type:int;
+                [Domain=<domain>; Path=<path>; HttpOnly; SameSite=Lax; Max-Age=<seconds>; expires=<datetime>]
 
+Web browsers and libraries for HTTP requests handling should automatically detect that header and register the defined
+``Cookie`` for subsequent requests. Alternatively, that ``Cookie`` can be provided directly in the request using the
+same format.
 
-All additional parameters are optional and can be provided to refine control of the scope the `Magpie` cookie applies
-to, notably to avoid conflicts with other potential cookies employed by the request. The only mandatory parts are the
-``MAGPIE_COOKIE_NAME`` value, the actual token value, and the indication ``!userid_type:int`` that tells `Magpie` the
-provided token information is employed to resolve the :term:`Logged User` by ID.
+All additional parameters (in brackets above) are optional and can be provided to explain how control of the scope the
+`Magpie` cookie applies to, notably to avoid conflicts with other potential cookies employed by the request. The only
+mandatory parts are the ``MAGPIE_COOKIE_NAME`` value, the actual token value, and the indication ``!userid_type:int``
+that tells `Magpie` the provided token information is employed to resolve the :term:`Logged User` by ID.
+
+Note that any ``Cookie`` generated by `Magpie` can have a maximum valid duration, identified by the both the returned
+``Max-Age`` in seconds and the ``expires`` value as explicit date and time. Generated cookies are defined in such a way
+that they will automatically emit a new ``Cookie`` based on reissue time after 1/10th of the ``Max-Age`` to update the
+``Cookie`` over continuous sessions. Any :term:`Logged User` will therefore remain logged in if further requests are
+accomplished using the same ``Cookie`` within the lifetime duration of the original login, unless explicitly logged out.
+Modifications of the duration is accomplished using configuration detailed in :ref:`config_security`.
+
+.. versionchanged:: 3.8
+    Although maximum duration could be defined in settings, prior versions did not explicitly indicate them in the
+    generated ``Cookie``. Following versions without these values will effectively mean the ``Cookie`` has unlimited
+    lifetime.
+
 
 As for most of the other API request endpoints offered by `Magpie`, the ``Accept`` header can be provided to select the
 format of the desired returned content. Following valid :term:`Authentication`, the body should contain a basic message
